pyproject.toml
[project]
name = "playlist-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andy Russell", email = "arussell123@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "google-api-python-client>=2.179.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "platformdirs>=4.4.0",
    "tqdm>=4.67.1",
]

[project.scripts]
playlist-scheduler = "playlist_scheduler:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
